0630 업데이트
- 프로그램 최적화
- 캐시 폴더 지정할 수 있도록 추가
- 업데이트 필수 조건 삭제
- 클립보드 확인 추가
- 구간 다운로드 추가
- 작업 취소 버튼 비활성화

+++ tmi_downloader.go
@@ -419,20 +419,24 @@ func main() { // 메인
 
 		w = a.NewWindow(title)
 		w.SetMainMenu(mainMenu)
-		w.CenterOnScreen()
-		w.SetFixedSize(false)
+		w.SetFixedSize(true)
 		w.SetIcon(appInfo.icon)
 		w.SetTitle(title)
 		w.SetContent(mainContent)
 		w.SetMaster()
 		w.Resize(fyne.NewSize(500, 550))
+		w.CenterOnScreen()
 
 		w.SetOnClosed(func() { // 강제 종료
 			if len(queue) != 0 {
 				fmt.Println("Quit: Forced")
 				for _, cmdProgress := range queue {
-					err = cmdProgress.CMD.Process.Kill()
-					ErrHandle(err)
+					if !cmdProgress.Done {
+						err = cmdProgress.CMD.Process.Kill()
+						if err != nil {
+							continue
+						}
+					}
 				}
 			}
